requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
gunicorn==23.0.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
alembic==1.16.5
python-dotenv==1.1.1
pydantic==2.11.7
requests==2.32.3
beautifulsoup4==4.12.3
